version

+++ src/main/java/com/backend/tienda/repository/Usuario_generalRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.backend.tienda.entity.Usuario_general;
 
-public interface Usuario_generalRepository extends JpaRepository<Usuario_general,Long>{
+public interface Usuario_generalRepository extends JpaRepository<Usuario_general,Integer>{
 	
 
 }
